Reorganize some stuff, handle update cancellation in optional flow
